Compute backoff before draining response body
Some APIs include details in the response body which is useful for
computing the correct backoff. Prior to this commit the body would be
drained before calling the backoff function, which meant that the
backoff function would not have access to the response body. This
confused me to a greater extent than I wish to admit.

This commit moves the drain after the backoff function.